+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * fix a bug where file descriptors weren't being cleaned up on a failed TCP
   connect
 * add h.consul:spawn, to facilitate spawning consul instances for testing
+* add a broadcast message primitive
 
 ## 0.3.1 - 2015-12-03